B 46 Refactor for QlabelRightClickMenu
